Edit to School21 version
Заменил все файлы на те варианты, которые будут сдаваться на проверку в Школе 21
