--- .github/PULL_REQUEST_TEMPLATE.md
@@ -1,10 +1,0 @@
-# Pull Request
-
-## Description of Changes
-Describe what has been changed and why. Specify which issues this PR addresses.
-
-## How to Test
-Describe how to test your changes. Include steps to follow to ensure everything works correctly.
-
-## Additional Notes
-Add any additional information that may be helpful for reviewers.
