--- .github/workflows/makefile.yml
@@ -1,23 +1,0 @@
-name: Makefile CI
-permissions:
-  contents: read
-
-on:
-  push:
-    branches: [ "main" ]
-  pull_request:
-    branches: [ "main" ]
-
-jobs:
-  build:
-
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/checkout@v4
-
-    - name: Build s21_cat
-      run: make
-
-    - name: Run tests
-      run: make check
