--- CONTRIBUTING.md
@@ -1,1 +1,0 @@
-Do whatever you want. But the code must work and be safe
